% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mcf7Rept}
\alias{mcf7Rept}
\title{Replication timing data}
\format{
\subsection{\code{mcf7Rept}}{

A \code{data.table} with 133,719 rows and 4 columns:
\describe{
\item{chr:}{Chromosome name}
\item{start:}{start positions of genomic ranges (open)}
\item{end:}{end positions of genomic ranges (closed)}
\item{signal:}{replication timing signal}
}
}
}
\source{
\url{https://genome-euro.ucsc.edu/cgi-bin/hgTables?db=hg19&hgta_group=regulation&hgta_track=wgEncodeUwRepliSeq&hgta_table=wgEncodeUwRepliSeqMcf7WaveSignalRep1&hgta_doSchema=describe+table+schema}
}
\usage{
mcf7Rept
}
\description{
Chromosome 10 Wavelet-smoothed signal data from MCF7 cells as provided by the UCSC genome browser (see source)
}
\keyword{datasets}
