% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptDir.r
\name{reptDir}
\alias{reptDir}
\title{reptDir: Replication Direction Finder}
\usage{
reptDir(reptdt, minLen, minSlope)
}
\arguments{
\item{reptdt}{A \code{data.table} with at least the following columns: \cr
\code{start}: start positions of genomic ranges (open) \cr
\code{end}: end positions of genomic ranges (closed) \cr
\code{signal}: replication time signal. \cr
The input \code{data.table} is assumed to have only genomic ranges
of a single chromosome ordered by start position.
The ranges are also assumed to be completely disjoint from one
another. The replication signal is assumed to be processed
already such that it meaningfully depicts the replication time profile of interest.
Such data is tipically found already in this format in the
UCSC genome browser for example. See the README at
\url{https://github.com/MikeACG/reptTools} for a full usage example.}

\item{minLen}{A scalar integer, minimum number of consecutive bases of direction domains.}

\item{minSlope}{A scalar numeric, minimum slope at both sides of ranges to be included in direction domains.}
}
\value{
A \code{data.table} with the following columns in addition to input columns: \cr
\code{dl}: signal difference between current range and previous range to the left \cr
\code{dr}: signal difference between next range to the right and current \cr
\code{extrema}: whether or not the range is a replication peak or valley \cr
\code{sdl}: sign of \code{dl} \cr
\code{sdr}: sign of \code{dr} \cr
\code{direction}: replication direction.
}
\description{
Determine the replication direction of genomic ranges based on replication timing signal.
}
\details{
The method used for assigning replication direction to each range
is inspired by the descriptions of \insertCite{Morganella2016_repdir;textual}{reptDir}.
Briefly, replication time signal extrema are found by identifying the first genomic ranges switching their slope
from positive to negative (peaks) or visceversa (valleys) after a streak of consecutive ranges with the same slope sign. Slopes
are defined as the difference between the signal
of any range and the previous neighbor range as defined by standard genomic coordinates in the reference strand.
Ranges with 0 slope and whose next neighbor also has 0 slope (flat regions), as well as
\code{extrema} themselves, are assigned \code{NA} replication direction.
Ranges for which a putative direction assignment is possible are
therefore those between extrema or between extrema and flat regions.
These collections of consecutive ranges are deemed replication direction domains.
Additional ranges may be assigned \code{NA} direction by \code{minLen}
and then by \code{minSlope} criteria in that order. These parameters
correspond to filters applied by \insertCite{Morganella2016_repdir;textual}{reptDir} and
\insertCite{Haradhvala2016_repdir;textual}{reptDir} respectively. if a domain spans less than
\code{minLen} nucleotides, its ranges are determined to have unreliable replication direction.
If a range or its next neighbor has a slope less than \code{minSlope},
its replication direction is deemed unreliable.
The remaining domains with only reliable ranges are assigned replication direction
based on the peaks being replication origins and valleys termination zones.
Therefore, a domain whose ranges have a positive slope indicates replication
from peak to valley in the "q-to-p" direction relative to the chromosomal arms.
A negative slope indicates peak to valley replication in the "p-to-q" direction.
The leading strand is the reference strand in q-to-p replication while
the lagging strand is the reference in the p-to-q case.
}
\references{
\insertAllCited{}
}
